requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.13.0
Pillow==10.0.1
numpy==1.24.3
pandas==1.5.3
